feat: split out scope into separate module
Further improve modularization in the auth area by splitting out the
AuthTokenScope into separate module and deduplicate errors by improving
error nesting.
