+++ openstack_sdk/src/auth/authtoken.rs
@@ -23,142 +23,129 @@ use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use tracing::{error, trace};
 
-//use crate::api::identity::v3::auth::os_federation::{
-//    identity_provider::protocol::websso::get as fed_idp_sso_get, websso::get as fed_sso_get,
-//};
 use crate::api::identity::v3::auth::token::create as token_v3;
 use crate::api::identity::v3::auth::token::get as token_v3_info;
 use crate::api::RestEndpoint;
-use crate::auth::{v3applicationcredential, v3password, v3token, v3totp, v3websso, AuthState};
+use crate::auth::{
+    authtoken_scope, v3applicationcredential, v3password, v3token, v3totp, v3websso, AuthState,
+};
 use crate::config;
-use crate::types::identity::v3::{AuthReceiptResponse, AuthResponse, Domain, Project};
+use crate::types::identity::v3::{AuthReceiptResponse, AuthResponse};
+
+pub use crate::auth::authtoken::authtoken_scope::AuthTokenScope;
 
 /// AuthToken (X-Auth-Token) based auth errors
 #[derive(Debug, Error)]
 #[non_exhaustive]
 pub enum AuthTokenError {
+    /// Header error
     #[error("header value error: {}", source)]
     HeaderValue {
         #[from]
         source: http::header::InvalidHeaderValue,
     },
 
-    #[error("Invalid auth_url: {}", source)]
-    InvalidAuthUrl {
-        #[from]
-        source: url::ParseError,
-    },
-
-    #[error("Cannot construct auth from config: {}", msg)]
-    Config { msg: String },
-
+    /// Unsupported identity method
     #[error(
         "AuthType `{}` is not a supported type for authenticating towards the cloud",
         auth_type
     )]
     IdentityMethod { auth_type: String },
 
+    /// Unsupported identity method in sync mode
     #[error(
         "AuthType `{}` is not a supported type for authenticating towards the cloud with sync interface",
         auth_type
     )]
     IdentityMethodSync { auth_type: String },
 
-    #[error("Cannot determine authorization scope from config")]
-    MissingScope,
+    /// Auth data is missing
     #[error("Auth data is missing")]
     MissingAuthData,
+
+    /// auth_url is missing
     #[error("Auth URL is missing")]
     MissingAuthUrl,
+
+    /// `Identity` request part build error
     #[error("Cannot construct identity auth information from config: {}", source)]
     AuthRequestIdentity {
+        /// The error source
         #[from]
         source: token_v3::IdentityBuilderError,
     },
-    #[error("Cannot construct project scope information from config: {}", source)]
-    AuthRequestScopeProject {
-        #[from]
-        source: token_v3::ProjectBuilderError,
-    },
-    #[error(
-        "Cannot construct project scope domain information from config: {}",
-        source
-    )]
-    AuthRequestScopeProjectDomain {
-        #[from]
-        source: token_v3::ProjectDomainBuilderError,
-    },
-    #[error(
-        "Cannot construct project scope domain information from config: {}",
-        source
-    )]
-    AuthRequestScopeDomain {
-        #[from]
-        source: token_v3::ScopeDomainBuilderError,
-    },
-    #[error("Cannot construct auth scope information from config: {}", source)]
-    AuthRequestScope {
-        #[from]
-        source: token_v3::ScopeBuilderError,
-    },
+
+    /// `Auth` request part build error
     #[error("error preparing auth request: {}", source)]
     AuthRequestAuth {
+        /// The error source
         #[from]
         source: token_v3::AuthBuilderError,
     },
+
+    /// Auth request build error
     #[error("error preparing auth request: {}", source)]
     AuthRequest {
+        /// The error source
         #[from]
         source: token_v3::RequestBuilderError,
     },
+
+    /// Token `Info` request build error
     #[error("error preparing token info request: {}", source)]
     InfoRequest {
+        /// The error source
         #[from]
         source: token_v3_info::RequestBuilderError,
+    },
+
+    /// Token Scope error
+    #[error("Scope error: {}", source)]
+    Scope {
+        /// The error source
+        #[from]
+        source: authtoken_scope::AuthTokenScopeError,
     },
 
     /// ApplicationCredentials Identity error
     #[error("ApplicationCredential authentication error: {}", source)]
     ApplicationCredential {
+        /// The error source
         #[from]
-        /// Error source
         source: v3applicationcredential::ApplicationCredentialError,
     },
 
     /// Password Identity error
     #[error("Password based authentication error: {}", source)]
     Password {
+        /// The error source
         #[from]
-        /// Error source
         source: v3password::PasswordError,
     },
 
     /// Token Identity error
     #[error("Token based authentication error: {}", source)]
     Token {
+        /// The error source
         #[from]
-        /// Error source
         source: v3token::TokenError,
     },
 
     /// TOTP Idetinty error
     #[error("Password based authentication error: {}", source)]
     Totp {
+        /// The error source
         #[from]
-        /// Error source
         source: v3totp::TotpError,
     },
 
     /// WebSSO Identity error
     #[error("SSO based authentication error: {}", source)]
     WebSso {
+        /// The error source
         #[from]
-        /// Error source
         source: v3websso::WebSsoError,
     },
-
-    #[error(transparent)]
-    Other(#[from] anyhow::Error),
 }
 
 type AuthResult<T> = Result<T, AuthTokenError>;
@@ -166,7 +153,9 @@ type AuthResult<T> = Result<T, AuthTokenError>;
 /// OpenStack AuthToken authorization structure
 #[derive(Clone, Default, Deserialize, Serialize)]
 pub struct AuthToken {
+    /// Token itself
     pub(crate) token: String,
+    /// Auth info reported by the server
     pub(crate) auth_info: Option<AuthResponse>,
 }
 
@@ -208,20 +197,12 @@ impl AuthToken {
     }
 
     /// Get Token scope information
-    pub fn get_scope(&self) -> AuthorizationScope {
+    pub fn get_scope(&self) -> AuthTokenScope {
         match &self.auth_info {
-            Some(ref data) => AuthorizationScope::from(data),
-            _ => AuthorizationScope::Unscoped,
+            Some(ref data) => AuthTokenScope::from(data),
+            _ => AuthTokenScope::Unscoped,
         }
     }
-}
-
-/// Represent authorization scope
-#[derive(Clone, Deserialize, Eq, Hash, PartialEq, Serialize, Debug)]
-pub enum AuthorizationScope {
-    Project(Project),
-    Domain(Domain),
-    Unscoped,
 }
 
 /// Supported AuthTypes
@@ -359,13 +340,13 @@ pub(crate) fn build_identity_data_from_config<'a>(
 /// Build Auth request from `Identity` and `AuthScope`
 pub(crate) fn build_auth_request_with_identity_and_scope<'a>(
     auth: &token_v3::Identity<'a>,
-    scope: &AuthorizationScope,
+    scope: &AuthTokenScope,
 ) -> Result<token_v3::Request<'a>, AuthTokenError> {
     let mut auth_request_data = token_v3::AuthBuilder::default();
     auth_request_data.identity(auth.clone());
     match scope {
         // For no scope we should not fill anything
-        AuthorizationScope::Unscoped => {}
+        AuthTokenScope::Unscoped => {}
         _ => {
             if let Ok(scope_data) = token_v3::Scope::try_from(scope) {
                 auth_request_data.scope(scope_data);
@@ -381,14 +362,14 @@ pub(crate) fn build_auth_request_with_identity_and_scope<'a>(
 /// Build Auth request from `AuthToken` and `AuthScope
 pub(crate) fn build_reauth_request<'a>(
     auth: &AuthToken,
-    scope: &AuthorizationScope,
+    scope: &AuthTokenScope,
 ) -> Result<token_v3::Request<'a>, AuthTokenError> {
     let identity_data = token_v3::Identity::try_from(auth)?;
     let mut auth_request_data = token_v3::AuthBuilder::default();
     auth_request_data.identity(identity_data);
     match scope {
         // For no scope we should not fill anything
-        AuthorizationScope::Unscoped => {}
+        AuthTokenScope::Unscoped => {}
         _ => {
             if let Ok(scope_data) = token_v3::Scope::try_from(scope) {
                 auth_request_data.scope(scope_data);
@@ -406,7 +387,7 @@ pub(crate) fn build_auth_request_from_receipt<'a>(
     config: &config::CloudConfig,
     receipt_header: HeaderValue,
     receipt_data: &AuthReceiptResponse,
-    scope: &AuthorizationScope,
+    scope: &AuthTokenScope,
     interactive: bool,
 ) -> Result<impl RestEndpoint + 'a, AuthTokenError> {
     let mut identity_builder = token_v3::IdentityBuilder::default();
@@ -469,13 +450,11 @@ pub(crate) fn build_token_info_endpoint<'a, S: AsRef<str>>(
 
 #[cfg(test)]
 mod tests {
-
     use serde_json::json;
 
     use super::*;
 
     use crate::config;
-    use crate::types::identity::v3::{AuthResponse, AuthToken};
 
     #[test]
     fn test_config_into_auth_password() -> Result<(), &'static str> {
@@ -537,39 +516,5 @@ mod tests {
             serde_json::to_value(auth_data).unwrap()
         );
         Ok(())
-    }
-
-    #[test]
-    fn test_auth_validity_unset() {
-        let auth = super::AuthToken::default();
-        assert!(matches!(auth.get_state(), AuthState::Unset));
-    }
-
-    #[test]
-    fn test_auth_validity_expired() {
-        let auth = super::AuthToken {
-            token: "".to_string(),
-            auth_info: Some(AuthResponse {
-                token: AuthToken {
-                    expires_at: chrono::offset::Local::now() - chrono::Duration::days(1),
-                    ..Default::default()
-                },
-            }),
-        };
-        assert!(matches!(auth.get_state(), AuthState::Expired));
-    }
-
-    #[test]
-    fn test_auth_validity_valid() {
-        let auth = super::AuthToken {
-            token: "".to_string(),
-            auth_info: Some(AuthResponse {
-                token: AuthToken {
-                    expires_at: chrono::offset::Local::now() + chrono::Duration::days(1),
-                    ..Default::default()
-                },
-            }),
-        };
-        assert!(matches!(auth.get_state(), AuthState::Valid));
     }
 }
